refactor(app): centralize preview JSON data
