+++ openspec/changes/refactor-app-preview-json-data/README.md
@@ -1,0 +1,3 @@
+# refactor-app-preview-json-data
+
+Refactor: centralize preview JSON data construction
