+++ src/app/mod.rs
@@ -3,6 +3,7 @@ pub(crate) mod doctor_next_actions;
 pub(crate) mod next_actions;
 pub(crate) mod operator_assets;
 pub(crate) mod preview_diff;
+pub(crate) mod preview_json;
 pub(crate) mod status_drift;
 pub(crate) mod status_json;
 pub(crate) mod status_next_actions;
